backend/pyproject.toml
[tool.poetry]
name = "datasmith"
version = "0.1.0"
description = ""
authors = ["Dantes <dantessvirid@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg-binary = "^3.2.3"
psycopg = {extras = ["binary"], version = "^3.2.3"}
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
dj-rest-auth = "^7.0.0"
djangorestframework-simplejwt = "^5.3.1"
pandasql = "^0.7.3"
django-extensions = "^3.2.3"
ipython = "^8.29.0"
ptpython = "^3.0.29"
mysql-connector-python = "^9.1.0"
boto3 = "^1.35.64"
django-storages = "^1.14.4"

scikit-learn = "^1.5.2"

cachetools = "^5.5.0"
django-cors-headers = "^4.6.0"
pymongo = "^4.6.1"
redis = "^5.0.0"
django-redis = "^5.4.0"

torch = "^2.7.1"
xgboost = "^3.0.2"
debugpy = "^1.8.14"
celery = "^5.5.3"
torchvision = "^0.22.1"
lightgbm = "^4.6.0"
channels = "^4.2.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"



[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
